third_party/blink/web_tests/fast/css/invalidation/selection-pseudo-expected.txt

Style invalidation for ::selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(select[0], '::selection').backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(select[0], '::selection').backgroundColor is green
PASS [object Object] is defined.
PASS getComputedStyle(select[1], '::selection').backgroundColor is red
PASS internals.updateStyleAndReturnAffectedElementCount() is 4
PASS getComputedStyle(select[1], '::selection').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE

1
2
